Move local control to a hai_agents_local sibling package outside the codegen surface
