+++ src/hai_agents_local/bridge.py
