+++ src/hai_agents_local/browser.py
