+++ src/hai_agents_local/desktop.py
