+++ src/hai_agents_local/errors.py
