+++ src/hai_agents_local/manager.py
